Add grayFilter and brigthnessFilter

+++ PhotoFilters.BL/Conversion.cs
@@ -30,7 +30,7 @@ namespace PhotoFilters.BL
 
             for (int i = 0; i < photo.Width; i++)
                 for (int j = 0; j < photo.Height; j++)
-                    bm.SetPixel(i, j, Color.FromArgb(photo[i, j].R, photo[i, j].G, photo[i, j].B));
+                    bm.SetPixel(i, j, Color.FromArgb((int)photo[i, j].R, (int)photo[i, j].G, (int)photo[i, j].B));
             return bm;
         }
     }
